re-organizing nav data restructuring
